v4.0.0-beta.173

+++ app/Actions/Database/StartMysql.php
@@ -56,7 +56,7 @@ class StartMysql
                     'memswap_limit' => $this->database->limits_memory_swap,
                     'mem_swappiness' => $this->database->limits_memory_swappiness,
                     'mem_reservation' => $this->database->limits_memory_reservation,
-                    'cpus' => (int) $this->database->limits_cpus,
+                    'cpus' => (float) $this->database->limits_cpus,
                     'cpuset' => $this->database->limits_cpuset,
                     'cpu_shares' => $this->database->limits_cpu_shares,
                 ]
